Extract state management
