+++ esp-wifi/src/wifi/os_adapter.rs
@@ -7,9 +7,7 @@
 pub(crate) mod os_adapter_chip_specific;
 
 use core::cell::RefCell;
-use core::sync::atomic::Ordering;
 
-use atomic_enum::atomic_enum;
 use critical_section::Mutex;
 use enumset::EnumSet;
 
@@ -34,57 +32,9 @@ use crate::compat::common::syslog;
 
 use super::WifiEvent;
 
-pub(crate) static STA_STATE: AtomicWifiState = AtomicWifiState::new(WifiState::Invalid);
-pub(crate) static AP_STATE: AtomicWifiState = AtomicWifiState::new(WifiState::Invalid);
-
 // useful for waiting for events - clear and wait for the event bit to be set again
 pub(crate) static WIFI_EVENTS: Mutex<RefCell<EnumSet<WifiEvent>>> =
     Mutex::new(RefCell::new(enumset::enum_set!()));
-
-#[atomic_enum]
-#[derive(PartialEq)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum WifiState {
-    StaStarted,
-    StaConnected,
-    StaDisconnected,
-    StaStopped,
-
-    ApStarted,
-    ApStopped,
-
-    Invalid,
-}
-
-impl From<WifiEvent> for WifiState {
-    fn from(event: WifiEvent) -> WifiState {
-        match event {
-            WifiEvent::StaStart => WifiState::StaStarted,
-            WifiEvent::StaConnected => WifiState::StaConnected,
-            WifiEvent::StaDisconnected => WifiState::StaDisconnected,
-            WifiEvent::StaStop => WifiState::StaStopped,
-            WifiEvent::ApStart => WifiState::ApStarted,
-            WifiEvent::ApStop => WifiState::ApStopped,
-            _ => WifiState::Invalid,
-        }
-    }
-}
-
-pub fn get_ap_state() -> WifiState {
-    AP_STATE.load(Ordering::Relaxed)
-}
-
-pub fn get_sta_state() -> WifiState {
-    STA_STATE.load(Ordering::Relaxed)
-}
-
-pub fn get_wifi_state() -> WifiState {
-    match super::get_wifi_mode() {
-        Ok(super::WifiMode::Sta) => get_sta_state(),
-        Ok(super::WifiMode::Ap) => get_ap_state(),
-        _ => WifiState::Invalid,
-    }
-}
 
 /****************************************************************************
  * Name: wifi_env_is_chip
@@ -934,16 +884,7 @@ pub unsafe extern "C" fn event_post(
     trace!("EVENT: {:?}", event);
     critical_section::with(|cs| WIFI_EVENTS.borrow_ref_mut(cs).insert(event));
 
-    match event {
-        WifiEvent::StaConnected
-        | WifiEvent::StaDisconnected
-        | WifiEvent::StaStart
-        | WifiEvent::StaStop => STA_STATE.store(WifiState::from(event), Ordering::Relaxed),
-        WifiEvent::ApStart | WifiEvent::ApStop => {
-            AP_STATE.store(WifiState::from(event), Ordering::Relaxed)
-        }
-        other => debug!("Unhandled event: {:?}", other),
-    }
+    super::state::update_state(event);
 
     #[cfg(feature = "async")]
     event.waker().wake();
