+++ esp-wifi/src/wifi/state.rs
@@ -1,0 +1,77 @@
+use super::WifiEvent;
+
+use atomic_enum::atomic_enum;
+use core::sync::atomic::Ordering;
+
+#[atomic_enum]
+#[derive(PartialEq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum WifiState {
+    StaStarted,
+    StaConnected,
+    StaDisconnected,
+    StaStopped,
+
+    ApStarted,
+    ApStopped,
+
+    Invalid,
+}
+
+impl From<WifiEvent> for WifiState {
+    fn from(event: WifiEvent) -> WifiState {
+        match event {
+            WifiEvent::StaStart => WifiState::StaStarted,
+            WifiEvent::StaConnected => WifiState::StaConnected,
+            WifiEvent::StaDisconnected => WifiState::StaDisconnected,
+            WifiEvent::StaStop => WifiState::StaStopped,
+            WifiEvent::ApStart => WifiState::ApStarted,
+            WifiEvent::ApStop => WifiState::ApStopped,
+            _ => WifiState::Invalid,
+        }
+    }
+}
+
+pub(crate) static STA_STATE: AtomicWifiState = AtomicWifiState::new(WifiState::Invalid);
+pub(crate) static AP_STATE: AtomicWifiState = AtomicWifiState::new(WifiState::Invalid);
+
+pub fn get_ap_state() -> WifiState {
+    AP_STATE.load(Ordering::Relaxed)
+}
+
+pub fn get_sta_state() -> WifiState {
+    STA_STATE.load(Ordering::Relaxed)
+}
+
+pub(crate) fn update_state(event: WifiEvent) {
+    match event {
+        WifiEvent::StaConnected
+        | WifiEvent::StaDisconnected
+        | WifiEvent::StaStart
+        | WifiEvent::StaStop => STA_STATE.store(WifiState::from(event), Ordering::Relaxed),
+
+        WifiEvent::ApStart | WifiEvent::ApStop => {
+            AP_STATE.store(WifiState::from(event), Ordering::Relaxed)
+        }
+
+        other => debug!("Unhandled event: {:?}", other),
+    }
+}
+
+#[cfg(feature = "async")]
+pub(crate) fn reset_ap_state() {
+    AP_STATE.store(WifiState::Invalid, Ordering::Relaxed)
+}
+
+#[cfg(feature = "async")]
+pub(crate) fn reset_sta_state() {
+    STA_STATE.store(WifiState::Invalid, Ordering::Relaxed)
+}
+
+pub fn get_wifi_state() -> WifiState {
+    match super::get_wifi_mode() {
+        Ok(super::WifiMode::Sta) => get_sta_state(),
+        Ok(super::WifiMode::Ap) => get_ap_state(),
+        _ => WifiState::Invalid,
+    }
+}
